Make it possible to drag the scene around with the mouse.
It is now possible to move the camera by clicking the left mouse button and dragging the mouse. Also, I refactored the camera move* and turn* methods a bit, so that a factor can be taken into account. This is used when responding to a move or turn caused by a mouse event. Finally, turning is now clamped or modulo'd (in a nice(r) way) in the Camera class, to prevent "flipping" of the camera. I tweaked the constants a bit and I think the camera movement is pretty fluid and intuitive right now. We might make it possible to change these constants in the settings though.

Side note: when moving the camera by clicking the left mouse button and dragging the mouse, the zoom is taken into account. This ensures that when you zoom in, you make a smaller movement when dragging the mouse the same number of pixels. I think this is intuitive. This is not (yet?) implemented for moving the camera using buttons. Not sure if that is necessary.