Implemented MenuBar, a class that can display a menu bar in the window.
A MenuBar is populated with MenuBarItem instances, which have text and an icon. All this stuff is created in ACCGProgram. Note that also the event handlers for onClick events are created there. I have taken icons from Oxygen (we need credits I think?) which I rescaled to make the size a power of 2 (namely 64). Also, I changed the mode from grayscale to RGB for the media-playback-* images because LWJGL did not support it. The MenuBar can be placed at the top, bottom, left or right of the screen. It can also be placed at the beginning, center or ending of one of these sides. We may make this a setting. To do so, we might need a "child" MenuBar, which is -- I think -- not very hard to implement. One could give a MenuBar a parent, which can be null, and draw it next to its parent if there is any. Just introduces more case distinctions...