Cleaned up some APIs. Made it possible to change the mouse sensitivity now.
The MenuBar can now indicate if it actually handled an event or not. This way, a mouse move event can be stopped from "bubbling up" to the camera. The same holds for scroll events, that can now be handled by the MenuBar as well. This is used in the new SliderMenuBarItem, which can be interacted with by clicking it and scrolling then. Some finetuning may be needed, but the initial version works.