Cleanup, factory methods, made it possible to give a menu item a type.
A MenuBarItem of type CHECKABLE or CHECKABLE_UNIQUE can be checked, much like a checkable QAction from Qt. The unique part means that other menu items in the same MenuBar are automatically unchecked when some MenuBarItem is checked. This is applicable to and useful for the menu position and alignment menus.