It is now possible to drag on sliders to change their value.
It is still possible to use the mouse wheel to change the value, but dragging is more intuitive. I made the bar a bit higher to make it easier to drag inside it (easier "aiming"). As a result of this, the API for MenuBarItems has a new member: onDrag(int x, int y).