The presentation of menu items can now be changed (also in the program).
There is a choice between what we had already (big icons above the text) and small icons left of the text. The choice is, just like alignment and position, stored in a Preferences object and is hence persistent. I personally must admit that I like the new choice better, but for example for the presentation, it may be more clear to use the old presentation style.