BendLeftConveyorBlock and BendRightConveyorBlock are animated properly.
I had to fix something in the normals, because of the ugly property that all points of one side are the same in a bend.
We could maybe draw the conveyor "nicely" in case it is connected to other conveyor blocks on both sides... That is,
with just a curve instead of first a small piece straight, then a curve and then straight again.