Fixed a LOT of bugs in the layouting code, initial work on relaying events to the GUI
Also, built the "real" toolbar (instead of the test that was there before) in the new GUI system. Also, made the SliderMenuBarItem compile and (hopefully) work when the events are relayed properly. Also, reverted the y-axis to point downwards, like in Swing. Also, probably a lot more small things that I forgot :)