Oops. Previous commit did not work... fixed now!
I actually removed the 'slow' Luggage-removing code now and fixed some bugs in the JBullet ContactProcessedCallback. Luggage is now removed as soon as it hits the floor in any way. This also fixes the bug where Luggage that stood on one side on the floor (not the top or bottom) would not be removed.