It is now no longer possible to place a block on/in another block.
The height of other blocks is now actually used to verify this. Also, the ShadowBlock can become 'alerted' to indicate that placement at that position is not possible. To ensure that the ShadowBlock is visible even when drawn over other blocks, a scale factor is introduced and the ShadowBlock is drawn slightly bigger when alerted.