Fixed some small see-through issues at back/front of Enter-/LeaveBlocks.
You could see a small part of the conveyor belt sticking out through the wall. Not anymore. The FlatConveyorBlock does have a notion of "being an Enter-/LeaveBlock" however, which may not be the "nicest" solution. We do have minimal code duplication now, however. And I think that it actually is a nice solution :-)