Added an additional level; furthermore removed the block limit from the existing levels
Rationale: I think that having a block limit on the start levels is only a hindrance for people exploring the program; as seen on the demo day, everyone likes to try out. I think that it is better to use the block limit as an additional challenge to make levels harder. :)